Migrated Spring OpenBank to Spring Boot
Refactored existing Dockerfile to be environment based,
requiring dependicy tools and kits to be installed on the environment
to be able to create a Docker trimmed down Wildfly image.

Refactored old pipeline Docker file to be to be more slimmed
and non-dependent environment trimmed Wildfly image.

Added pipeline and ordinary Docker Compose files to ease up the
execution of the OpenBank container.

Made the component/service/repository constructors public.

Refactored persistence.xml properties to be set instead from a
property fields file and imported by the EntityManagerFactoryBean.

Refactored Broker service and updated the JmsMessageQueueSender
system test.

Refactored thew Docker files to be able to execute the migrated
Spring Boot application.

Updated REST API documentation for the new base URL.

Updated README.md file for running the ported Spring Boot
application.

+++ src/test/java/se/jsquad/repository/ClientRepositoryImplTest.java
@@ -42,13 +42,13 @@ public class ClientRepositoryImplTest {
         assertEquals("john.doe@test.se", client.getPerson().getMail());
         assertEquals(client, client.getPerson().getClient());
 
-        assertEquals(500, ((RegularClient) client.getClientType()).getRating());
+        assertEquals(500, ((RegularClient) client.getClientType()).getRating().longValue());
         assertEquals(client, client.getClientType().getClient());
 
         Account account = client.getAccountSet().iterator().next();
 
         assertEquals("1000", account.getAccountNumber());
-        assertEquals(500, account.getBalance());
+        assertEquals(500, account.getBalance().longValue());
         assertEquals(client, account.getClient());
 
         AccountTransaction accountTransaction = account.getAccountTransactionSet().iterator().next();
